requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
sqlalchemy>=2.0.25
python-dotenv>=1.0.0
pydantic>=2.6.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
email-validator>=2.1.0
pytest>=8.0.0
httpx>=0.26.0
websockets>=12.0
python-multipart>=0.0.6
openpyxl>=3.1.0
reportlab>=4.0.0